Refactor model handling: update ModelAdapter and ModelItem for clarity, rename SpintaModelRef to RowModelRef, and clean up imports in XML domain modules
