+++ spinta/datasets/backends/dataframe/backends/xml/adapter/dask_xml.py
@@ -3,8 +3,9 @@ from typing import Any, Dict, Iterable, Mapping, Optional
 
 import dask.dataframe as dd
 
+from spinta.adapters.loaders import DataAdapter, DataAdapterError
 from spinta.datasets.backends.dataframe.backends.xml.adapter.row import RowList
-from spinta.datasets.backends.dataframe.backends.xml.domain.model import DataAdapter, DataAdapterError, ModelHeader
+from spinta.datasets.backends.dataframe.backends.xml.domain.model import ModelHeader
 from spinta.utils.schema import NA
